+++ src/components/cart/CartListItem/CartListItem.styles.ts
@@ -1,51 +1,6 @@
-import QuantityInput from '../QuantityInput';
-import { CartItem } from '../../types';
 import { styled } from 'styled-components';
-import { BsFillTrash3Fill } from 'react-icons/bs';
-import { useProduct } from '../../hooks/useProduct';
-import { Dispatch, MouseEvent, SetStateAction } from 'react';
 
-interface Props {
-  item: CartItem;
-  setCheckItems: Dispatch<SetStateAction<number[]>>;
-}
-
-const CartListItem = ({ item, setCheckItems }: Props) => {
-  const {
-    removeItem,
-    handleNumberInputChange,
-    handleIncreaseItem,
-    handleDecreaseCartItem,
-  } = useProduct(item.product.id);
-
-  const handleRemoveFromCart = (id: number) => (e: MouseEvent<HTMLButtonElement>) => {
-    const confirmResult = window.confirm('정말로 삭제하시겠습니까?');
-    if (confirmResult) {
-      setCheckItems((prev) => prev.filter((itemId) => itemId !== id));
-      removeItem();
-    }
-  };
-
-  return (
-    <S.Wrapper>
-      <img src={item.product.imageUrl} alt={String(item.id)} />
-      <S.Name>{item.product.name}</S.Name>
-      <S.RemoveButton onClick={handleRemoveFromCart(item.id)}>
-        <BsFillTrash3Fill size={24} />
-      </S.RemoveButton>
-      <QuantityInput
-        id={String(item.id)}
-        value={item.quantity}
-        onChange={handleNumberInputChange}
-        onIncrement={handleIncreaseItem}
-        onDecrement={handleDecreaseCartItem}
-      />
-      <S.Price>{(item.product.price * item.quantity).toLocaleString()}원</S.Price>
-    </S.Wrapper>
-  );
-};
-
-const S = {
+export const S = {
   Wrapper: styled.div`
     display: flex;
     flex-direction: row;
@@ -144,5 +99,3 @@ const S = {
     }
   `,
 };
-
-export default CartListItem;
