+++ src/components/cart/TotalCheckbox/index.tsx
@@ -1,5 +1,4 @@
-import { styled } from 'styled-components';
-import { CheckboxStyle } from '../@common/Checkbox';
+import { S } from './TotalCheckbox.styles';
 
 interface Props {
   cartLength: number;
@@ -28,29 +27,6 @@ const TotalCheckbox = ({
       </S.RemoveSelectedButton>
     </S.CheckboxAllWrapper>
   );
-};
-
-const S = {
-  CheckboxAllWrapper: styled.div`
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    height: 40px;
-    gap: 10px;
-    font-size: 24px;
-  `,
-
-  RemoveSelectedButton: styled.button`
-    width: 100px;
-    height: 35px;
-    padding: 5px;
-    background-color: white;
-    border: 1px solid #bbb;
-    border-radius: 7px;
-    font-size: 16px;
-  `,
-
-  CheckboxAll: styled(CheckboxStyle)``,
 };
 
 export default TotalCheckbox;
